+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased
 
 - feat: Add trace header to http requests (#1213)
+- fix: ViewController swizzling before iOS 13 (#1231)
+- fix: AppStartMeasurement didFinishLaunching is nil (#1228)
+
+## 7.2.0-beta.5
+
+- perf: Scope sync to SentryCrash (#1193)
+- fix: Compile failure with C99 (#1224)
 
 ## 7.2.0-beta.4